+++ examples/04_data.rs
@@ -1,6 +1,6 @@
 use afire::{Header, Method, Query, Response, Server};
 
-// Send data to server with a Query String and Form Data
+// Send data to server with a Query String, Path Prams and Form Data
 
 fn main() {
     // Create a new Server instance on localhost port 8080
@@ -59,8 +59,32 @@ fn main() {
             .header(Header::new("Content-Type", "text/html"))
     });
 
+    // Define a page with path params
+    // As this is not built into afire you will need to use middleware (for now :P)
+    // This is like "/greet/:name"
+    server.middleware(Box::new(|req| {
+        // Move on if path dosent start with /path or the method is not get
+        // You could also do this with a regex (but afire is dependency free so I wont show that in this example)
+        if !req.path.starts_with("/greet/") || req.method != Method::GET {
+            return None;
+        }
+
+        // Extract the pram value
+        // This is getting the 2rd path value
+        // Ex: /greet/Darren/
+        let name = req.path.split("/").nth(2).unwrap();
+
+        // Respond with "Hello, {{NAME}}"
+        Some(
+            Response::new()
+                .text(format!("<h1>Hello, {}</h1>", name))
+                .header(Header::new("Content-Type", "text/html")),
+        )
+    }));
+
     // You can now goto http://localhost:8080?name=John and should see "Hello, John"
     // If you goto http://localhost:8080/form and submit the form you should see "Hello, {NAME}"
+    // Also goto http://localhost:8080/greet/John and you should see "Hello, John"
 
     println!(
         "[04] Listening on http://{}:{}",